Report on missing commands in update script (#38)
So far, we simply reported that a command failed in the dev. script for
updating to the latest meta-model. However, when the command does not
exit, an exception is raised and the traceback is reported, which is
uninformative for the user.

In this patch, we report the exact command which could not be found.